django/requirements.txt
django<4
django-cors-headers<4
django-import-export<3
django-rest-passwordreset<2
djangorestframework<4
djangorestframework-simplejwt<5
gunicorn<21
psycopg2<3
requests<3

django-environ==0.11.2

# GraphQL
graphene==3.3
django-graphql-jwt==0.4.0
graphene-django==3.2.2
openpyxl==3.1.5